refactor : SessionRegistration을 students 디렉토리로 옮긴다
