+++ src/main/java/nextstep/sessions/infrastructure/JdbcSessionRepository.java
@@ -10,7 +10,7 @@ import nextstep.sessions.domain.Session;
 import nextstep.sessions.domain.SessionBody;
 import nextstep.sessions.domain.SessionDate;
 import nextstep.sessions.domain.SessionProgressStatus;
-import nextstep.sessions.domain.SessionRegistration;
+import nextstep.sessions.domain.students.SessionRegistration;
 import nextstep.sessions.domain.SessionRepository;
 import nextstep.sessions.domain.SessionRecruitingStatus;
 import nextstep.sessions.domain.students.Student;
